requirements.txt
python-telegram-bot==20.3
redis==4.5.5
mysql-connector-python==8.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2

upgrade python-telegram-bot
